requirements.txt
tqdm
pyyaml
tiktoken
pypdf
python-docx
wikipedia
arxiv
chromadb
gradio
gradio-toggle
groq
nomic[local]
langchain
langchain-nomic
langchain-chroma
langchain-community
langchainhub
litellm
litellm[proxy]
ollama
openai
google-generativeai
duckduckgo-search